(backend)  replace dominant color with dominant color array
Migrates from a single dominant color string to an array of dominant color strings for both albums and artists.
This allows for a more nuanced representation of the prominent colors in thumbnails.

Removes the `dominant_color` column from the `ARTIST` and `ALBUM` tables.
Updates the data models, controllers, and services to use the new `thumbnailDominantColors` array instead of the single `dominantColor` string.

Updates the image service calls to get the top 3 dominant colors instead of just one.

delete unused imageService method