(frontend) add rater context
Implements a context panel to display and manage filters applied during the rating process.

This component allows users to view selected artists and albums, and toggle them to filter the displayed songs.

The RaterContext component uses the `useMusicStateAndDispatch` hook to access the current state and dispatch actions to update the rater filters.

CSS has been added to style the component and its entries.